Add x86 support
Adapted the scripts for x86. The scripts
automatically detect the host architecture
and build the corresponding image.

+++ image/build_image.sh
@@ -2,28 +2,30 @@
 set -x -e
 
 SCRIPTPATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
-
 PACKER_VERSION="1.11.2"
 
+# Disk image configs
+ARCH=$(dpkg --print-architecture)
+OS_VERSION="jammy"
 
-pushd $SCRIPTPATH
+
+cd $SCRIPTPATH
 
 if [ ! -f ./packer ]; then
-    wget https://releases.hashicorp.com/packer/${PACKER_VERSION}/packer_${PACKER_VERSION}_linux_arm64.zip;
-    unzip packer_${PACKER_VERSION}_linux_arm64.zip;
-    rm packer_${PACKER_VERSION}_linux_arm64.zip;
+    wget https://releases.hashicorp.com/packer/${PACKER_VERSION}/packer_${PACKER_VERSION}_linux_${ARCH}.zip;
+    unzip packer_${PACKER_VERSION}_linux_${ARCH}.zip;
+    rm packer_${PACKER_VERSION}_linux_${ARCH}.zip;
 fi
 
 
 # Store the configuration file name from the command line argument
-config_file=$SCRIPTPATH/ubuntu.pkr.hcl
+config_file=$SCRIPTPATH/ubuntu.${ARCH}.pkr.hcl
 
 
-if [ ! -f ./flash0.img ]; then
+if [ ${ARCH} == arm64 ] && [ ! -f ./flash0.img ] ; then
     cp /usr/share/qemu-efi-aarch64/QEMU_EFI.fd ./flash0.img
     truncate -s 64M flash0.img
 fi
-
 
 
 # Check if the specified configuration file exists
@@ -31,10 +33,10 @@ if [ -f "$config_file" ]; then
     # Install the needed plugins
     ./packer init "$config_file"
     # Build the image
-    ./packer build "$config_file"
+    ./packer build -var "ubuntu_version=${OS_VERSION}" "$config_file"
 else
     echo "Error: Configuration file '$config_file' not found."
     exit 1
 fi
 
-popd
+cd ..
